# server/src/api/tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TokenResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Emoji": {
				"properties": {
					"name": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"value",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserStatusValue": {
				"anyOf": [
					{
						"type": "string",
						"enum": [
							"online"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"offline"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"dnd"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"away"
						],
						"nullable": false
					}
				]
			},
			"UserStatus": {
				"properties": {
					"value": {
						"$ref": "#/components/schemas/UserStatusValue"
					},
					"statusText": {
						"type": "string"
					}
				},
				"required": [
					"value"
				],
				"type": "object"
			},
			"ApiUser": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/UserStatus"
					},
					"bot": {
						"type": "boolean"
					},
					"avatarUrl": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"bot",
					"avatarUrl",
					"name",
					"id"
				],
				"type": "object"
			},
			"ApiChannel": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/ApiUser"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateChannelRequest": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmbedAuthor": {
				"properties": {
					"iconUrl": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"iconUrl",
					"value"
				],
				"type": "object"
			},
			"EmbedPayload": {
				"properties": {
					"thumbnailUrl": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"footer": {
						"type": "string"
					},
					"color": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/EmbedAuthor"
					}
				},
				"type": "object"
			},
			"CreateMessageRequest": {
				"properties": {
					"content": {
						"type": "string"
					},
					"embed": {
						"$ref": "#/components/schemas/EmbedPayload"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "veld-chat-server",
		"version": "1.0.0",
		"description": "fun stuff happening",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/users/token": {
			"get": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/emojis": {
			"get": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Emoji"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/channels": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiChannel"
								}
							}
						}
					}
				},
				"security": [
					{
						"bot": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateChannelRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/channels/{channelId}/join": {
			"post": {
				"operationId": "JoinChannel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [
					{
						"bot": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "channelId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/channels/{channelId}/messages": {
			"post": {
				"operationId": "CreateMessage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"bot": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "channelId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMessageRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}